tests: update tests to ues new module
